Time: 205 ms (76.68%), Space: 76.4 MB (76.68%) - LeetHub

+++ 1833-maximum-ice-cream-bars/1833-maximum-ice-cream-bars.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int maxIceCream(vector<int>& a, int c) {
         sort(a.begin(), a.end());
-        long long cum = 0, ans = 0;
+        int cum = 0, ans = 0;
         for (auto x : a) {
             cum += x;
             if (cum > c) break;
